+++ tests/testthat/test-utilities-population.R
@@ -8,9 +8,16 @@ test_that("`sampleRandomValue()` generates needed distribution", {
   expect_equal(
     sampleRandomValue(Distributions$Normal, 5, 2, 10),
     c(
-      3.87904870689558, 4.53964502103344, 8.11741662829825, 5.14101678284915,
-      5.25857547032189, 8.43012997376656, 5.9218324119784, 2.46987753078693,
-      3.62629429621295, 4.10867605980008
+      3.87904870689558,
+      4.53964502103344,
+      8.11741662829825,
+      5.14101678284915,
+      5.25857547032189,
+      8.43012997376656,
+      5.9218324119784,
+      2.46987753078693,
+      3.62629429621295,
+      4.10867605980008
     ),
     tolerance = 0.001
   )
@@ -19,15 +26,24 @@ test_that("`sampleRandomValue()` generates needed distribution", {
   expect_equal(
     sampleRandomValue(Distributions$LogNormal, 5, 2, 10),
     c(
-      3.74081271106427, 4.24843764475839, 8.46318202896501, 4.77021554349172,
-      4.87946908411847, 8.98864517081978, 5.54444951200875, 2.85153959957418,
-      3.56304555191325, 3.90999158989997
+      3.74081271106427,
+      4.24843764475839,
+      8.46318202896501,
+      4.77021554349172,
+      4.87946908411847,
+      8.98864517081978,
+      5.54444951200875,
+      2.85153959957418,
+      3.56304555191325,
+      3.90999158989997
     ),
     tolerance = 0.001
   )
 })
 
 test_that("It creates population characteristics with ontogenies from excel", {
+  skip_on_os("mac")
+
   excelPath <- testConfigurationsPath("Populations.xlsx")
 
   populationCharacteristics <- readPopulationCharacteristicsFromXLS(
@@ -61,6 +77,8 @@ test_that("It creates population characteristics with ontogenies from excel", {
 })
 
 test_that("It creates population characteristics without ontogenies from excel", {
+  skip_on_os("mac")
+
   excelPath <- testConfigurationsPath("Populations.xlsx")
 
   populationCharachterstics <- readPopulationCharacteristicsFromXLS(
@@ -92,7 +110,11 @@ test_that("It creates population characteristics without ontogenies from excel",
 test_that("extendPopulationByUserDefinedParams works", {
   set.seed(42)
 
-  population <- ospsuite::loadPopulation(system.file("extdata", "SimResults_pop.csv", package = "ospsuite"))
+  population <- ospsuite::loadPopulation(system.file(
+    "extdata",
+    "SimResults_pop.csv",
+    package = "ospsuite"
+  ))
 
   esqlabsR::extendPopulationByUserDefinedParams(
     population = population,
@@ -114,32 +136,42 @@ test_that("extendPopulationFromXLS works", {
     code = {
       .writeExcel(
         path = PopulationParameters,
-        data = list("UserDefinedVariability" = data.frame(
-          `Container Path` = c("Organism|Kidney", "Organism|Kidney"),
-          `Parameter Name` = c("GFR", "eGFR"),
-          "Mean" = 0.12,
-          "SD" = 0.001,
-          "Distribution" = "Normal",
-          check.names = FALSE
-        ))
+        data = list(
+          "UserDefinedVariability" = data.frame(
+            `Container Path` = c("Organism|Kidney", "Organism|Kidney"),
+            `Parameter Name` = c("GFR", "eGFR"),
+            "Mean" = 0.12,
+            "SD" = 0.001,
+            "Distribution" = "Normal",
+            check.names = FALSE
+          )
+        )
       )
 
-      population <- ospsuite::loadPopulation(system.file("extdata", "SimResults_pop.csv", package = "ospsuite"))
+      population <- ospsuite::loadPopulation(system.file(
+        "extdata",
+        "SimResults_pop.csv",
+        package = "ospsuite"
+      ))
 
       set.seed(42)
-      extendPopulationFromXLS(population,
+      extendPopulationFromXLS(
+        population,
         PopulationParameters,
         sheet = "UserDefinedVariability"
       )
       expect_snapshot(
         population$getParameterValuesForIndividual(4)
       )
-      expect_equal(population$allParameterPaths, c(
-        "Organism|PeripheralVenousBlood|Aciclovir|Plasma (Peripheral Venous Blood)",
-        "Organism|Muscle|Intracellular|Aciclovir|Concentration",
-        "Organism|Kidney|GFR",
-        "Organism|Kidney|eGFR"
-      ))
+      expect_equal(
+        population$allParameterPaths,
+        c(
+          "Organism|PeripheralVenousBlood|Aciclovir|Plasma (Peripheral Venous Blood)",
+          "Organism|Muscle|Intracellular|Aciclovir|Concentration",
+          "Organism|Kidney|GFR",
+          "Organism|Kidney|eGFR"
+        )
+      )
     }
   )
 })
@@ -149,23 +181,29 @@ test_that("extendPopulationFromXLS throws an error if the sheet has wrong struct
     new = "PopulationParameters",
     fileext = ".xlsx",
     code = {
-      population <- ospsuite::loadPopulation(system.file("extdata", "SimResults_pop.csv", package = "ospsuite"))
+      population <- ospsuite::loadPopulation(system.file(
+        "extdata",
+        "SimResults_pop.csv",
+        package = "ospsuite"
+      ))
 
       .writeExcel(
         path = PopulationParameters,
-        data = list("UserDefinedVariability" = data.frame(
-          `Container Path` = character(),
-          `Parameter Name` = character(),
-          "Mean" = numeric(),
-          "SD" = numeric(),
-          # "Distribution" = character(),  # Distribution column is missing
-          check.names = FALSE
-        ))
+        data = list(
+          "UserDefinedVariability" = data.frame(
+            `Container Path` = character(),
+            `Parameter Name` = character(),
+            "Mean" = numeric(),
+            "SD" = numeric(),
+            # "Distribution" = character(),  # Distribution column is missing
+            check.names = FALSE
+          )
+        )
       )
 
-
       expect_error(
-        extendPopulationFromXLS(population,
+        extendPopulationFromXLS(
+          population,
           PopulationParameters,
           sheet = "UserDefinedVariability"
         ),
@@ -174,18 +212,21 @@ test_that("extendPopulationFromXLS throws an error if the sheet has wrong struct
 
       .writeExcel(
         path = PopulationParameters,
-        data = list("UserDefinedVariability" = data.frame(
-          "Container.Path" = character(), # column name is wrong
-          `Parameter Name` = character(),
-          "Mean" = numeric(),
-          "SD" = numeric(),
-          "Distribution" = character(),
-          check.names = FALSE
-        ))
+        data = list(
+          "UserDefinedVariability" = data.frame(
+            "Container.Path" = character(), # column name is wrong
+            `Parameter Name` = character(),
+            "Mean" = numeric(),
+            "SD" = numeric(),
+            "Distribution" = character(),
+            check.names = FALSE
+          )
+        )
       )
 
       expect_error(
-        extendPopulationFromXLS(population,
+        extendPopulationFromXLS(
+          population,
           PopulationParameters,
           sheet = "UserDefinedVariability"
         ),
@@ -202,20 +243,27 @@ test_that("extendPopulationFromXLS throws an error if specified sheet is empty o
     code = {
       .writeExcel(
         path = PopulationParameters,
-        data = list("UserDefinedVariability" = data.frame(
-          `Container Path` = character(),
-          `Parameter Name` = character(),
-          "Mean" = numeric(),
-          "SD" = numeric(),
-          "Distribution" = character(), # Distribution column is missing
-          check.names = FALSE
-        ))
+        data = list(
+          "UserDefinedVariability" = data.frame(
+            `Container Path` = character(),
+            `Parameter Name` = character(),
+            "Mean" = numeric(),
+            "SD" = numeric(),
+            "Distribution" = character(), # Distribution column is missing
+            check.names = FALSE
+          )
+        )
       )
 
-      population <- ospsuite::loadPopulation(system.file("extdata", "SimResults_pop.csv", package = "ospsuite"))
+      population <- ospsuite::loadPopulation(system.file(
+        "extdata",
+        "SimResults_pop.csv",
+        package = "ospsuite"
+      ))
 
       expect_error(
-        extendPopulationFromXLS(population,
+        extendPopulationFromXLS(
+          population,
           PopulationParameters,
           sheet = "UserDefinedVariability"
         ),
@@ -224,19 +272,22 @@ test_that("extendPopulationFromXLS throws an error if specified sheet is empty o
 
       .writeExcel(
         path = PopulationParameters,
-        data = list("UserDefinedVariability" = data.frame(
-          `Container Path` = "Organism|Kidney",
-          `Parameter Name` = "GFR",
-          "Mean" = 0.12,
-          "SD" = 0.001,
-          "Distribution" = NA,
-          check.names = FALSE
-        ))
+        data = list(
+          "UserDefinedVariability" = data.frame(
+            `Container Path` = "Organism|Kidney",
+            `Parameter Name` = "GFR",
+            "Mean" = 0.12,
+            "SD" = 0.001,
+            "Distribution" = NA,
+            check.names = FALSE
+          )
+        )
       )
 
       expect_snapshot(
         error = TRUE,
-        extendPopulationFromXLS(population,
+        extendPopulationFromXLS(
+          population,
           PopulationParameters,
           sheet = "UserDefinedVariability"
         )