+++ NEWS.md
@@ -4,7 +4,9 @@
 
 - User-defined parameters passed to the `createScenarios()` or `Scenario$new()` 
 in the `customParams` argument are applied last. Up to this version, they 
-were overwritten by the administration protocol (\#817)
+were overwritten by the administration protocol (\#817).
+- Added `snapshotProjectConfiguration()` and `restoreProjectConfiguration()` functions 
+to support version control and project sharing of project configurations.
 
 ## Minor improvements and bug fixes